+++ src/client/app-render.ts
@@ -389,7 +389,7 @@ export default class CanvasRenderer extends EventEmitter {
         this.plateBase.scaleDisplayZ = this.plateDisplay!.scaling.z
         this.plateBase.rotationZ     = this.plate!.rotation.z
         this.plateBase.positionZ     = this.plate!.position.z
-        this.plateBase.positionX     = this.plate!.position.x
+        this.plateBase.positionX     = this.plateDisplay!.position.x
 
         /*  gather references to hologram mesh nodes  */
         this.hologram        = this.scene.getNodeByName("Hologram")         as BABYLON.Nullable<BABYLON.TransformNode>
@@ -405,7 +405,7 @@ export default class CanvasRenderer extends EventEmitter {
         this.hologramBase.scaleDisplayZ = this.hologramDisplay!.scaling.z
         this.hologramBase.rotationZ     = this.hologram!.rotation.z
         this.hologramBase.positionZ     = this.hologram!.position.z
-        this.hologramBase.positionX     = this.hologram!.position.x
+        this.hologramBase.positionX     = this.hologramDisplay!.position.x
 
         /*  setup light shadow casting the display onto the wall  */
         const setupLight = (light: BABYLON.PointLight) => {
@@ -1254,7 +1254,7 @@ export default class CanvasRenderer extends EventEmitter {
             if (state.plate.lift !== undefined)
                 this.plate.position.z = this.plateBase.positionZ + state.plate.lift
             if (state.plate.distance !== undefined)
-                this.plate.position.x = this.plateBase.positionX - state.plate.distance
+                this.plateDisplay.position.x = this.plateBase.positionX - state.plate.distance
             if (state.plate.fadeTime !== undefined && this.plateFade !== state.plate.fadeTime)
                 this.plateFade = state.plate.fadeTime
             if (state.plate.opacity !== undefined) {
@@ -1407,7 +1407,7 @@ export default class CanvasRenderer extends EventEmitter {
             if (state.hologram.lift !== undefined)
                 this.hologram.position.z = this.hologramBase.positionZ + state.hologram.lift
             if (state.hologram.distance !== undefined)
-                this.hologram.position.x = this.hologramBase.positionX - state.hologram.distance
+                this.hologramDisplay.position.x = this.hologramBase.positionX - state.hologram.distance
             if (state.hologram.fadeTime !== undefined && this.hologramFade !== state.hologram.fadeTime)
                 this.hologramFade = state.hologram.fadeTime
             if (state.hologram.opacity !== undefined) {